+++ src/Order/Dao/Impl/OrderItemRefundDaoImpl.php
@@ -16,10 +16,23 @@ class OrderItemRefundDaoImpl extends GeneralDaoImpl implements OrderItemRefundDa
         ));
     }
 
+    public function findByOrderRefundIds($orderRefundIds)
+    {
+        return $this->findInField('order_refund_id', $orderRefundIds);
+    }
+
     public function declares()
     {
         return array(
-            'timestamps' => array('created_time', 'updated_time')
+            'timestamps' => array('created_time', 'updated_time'),
+            'orderbys' => array(
+                'id',
+                'created_time'
+            ),
+            'serializes' => array(
+            ),
+            'conditions' => array(
+            )
         );
     }
 }
