+++ src/Order/Dao/OrderItemDao.php
@@ -7,4 +7,6 @@ use Codeages\Biz\Framework\Dao\GeneralDaoInterface;
 interface OrderItemDao extends GeneralDaoInterface
 {
     public function findByOrderId($orderId);
+
+    public function findByOrderIds($orderIds);
 }
