+++ src/Order/Dao/OrderItemRefundDao.php
@@ -7,4 +7,6 @@ use Codeages\Biz\Framework\Dao\GeneralDaoInterface;
 interface OrderItemRefundDao extends GeneralDaoInterface
 {
     public function findByOrderRefundId($orderRefundId);
+
+    public function findByOrderRefundIds($orderRefundIds);
 }
