+++ src/Order/Service/Impl/OrderRefundServiceImpl.php
@@ -9,6 +9,11 @@ use Codeages\Biz\Framework\Util\ArrayToolkit;
 
 class OrderRefundServiceImpl extends BaseService implements OrderRefundService
 {
+    public function getById($id)
+    {
+        return $this->getOrderRefundDao()->get($id);
+    }
+    
     public function searchRefunds($conditions, $orderby, $start, $limit)
     {
         return $this->getOrderRefundDao()->search($conditions, $orderby, $start, $limit);
@@ -19,8 +24,28 @@ class OrderRefundServiceImpl extends BaseService implements OrderRefundService
         return $this->getOrderRefundDao()->count($conditions);
     }
 
+    public function searchRefundItems($conditions, $orderby, $start, $limit)
+    {
+        return $this->getOrderItemRefundDao()->search($conditions, $orderby, $start, $limit);
+    }
+
+    public function countRefundItems($conditions)
+    {
+        return $this->getOrderItemRefundDao()->count($conditions);
+    }
+
+    public function findRefundItemsByRefundIds($refundIds)
+    {
+        return $this->getOrderItemRefundDao()->findByOrderRefundIds($refundIds);
+    }
+
     protected function getOrderRefundDao()
     {
         return $this->biz->dao('Order:OrderRefundDao');
+    }
+
+    protected function getOrderItemRefundDao()
+    {
+        return $this->biz->dao('Order:OrderItemRefundDao');
     }
 }
